Code size optimization: shrink m_GfiRetryCnt and m_NoGndRetryCnt fields
There is a chance this optimization is incorrect, so let me know if so.

Since '255' is the "off" value for these checks, we can assume we'll
never need to track more than 255 retries for either the GFI or No
Ground checks. Thus, use an 8-bit variable for each.

This saves 6 bytes of RAM and 116 bytes of Flash.

+++ firmware/open_evse/J1772EvseController.h
@@ -114,7 +114,7 @@ class J1772EVSEController {
 #ifdef GFI
   Gfi m_Gfi;
   unsigned long m_GfiFaultStartMs;
-  unsigned long m_GfiRetryCnt;
+  uint8_t m_GfiRetryCnt;
   uint8_t m_GfiTripCnt; // contains tripcnt-1
 #endif // GFI
   AdcPin adcPilot;
@@ -148,7 +148,7 @@ class J1772EVSEController {
 #endif
 #ifdef ADVPWR
   unsigned long m_NoGndStart;
-  unsigned long m_NoGndRetryCnt;
+  uint8_t m_NoGndRetryCnt;
   uint8_t m_NoGndTripCnt; // contains tripcnt-1
   unsigned long m_StuckRelayStartTimeMS;
   uint8_t m_StuckRelayTripCnt; // contains tripcnt-1
